Fix Issue #115 bad kerning
It seems GDI+ renders already the LSB (left side bearing). So
there is no reason to pass it again to glyph.set_bearings.
The only case where it matters is when the LSB is negative. In
that case, the rect layout is pushed to the right so that the
whole glyph can be rendered correctly. But the pyglet still needs
to shift the glyph back to the left for correct kerning.