Fix issue #81 by removing circular references
There were several circular references in the Windows classes which
prevented them from being *immediately* garbage collected due to
reference counting going down to 0.

For Win32Window, there is a closure in _get_window_proc which is used
in 2 different places:
- 176: self._window_class.lpfnWndProc = WNDPROC(
                self._get_window_proc(self._event_handlers))
- 190: self._view_window_class.lpfnWndProc = WNDPROC(
                self._get_window_proc(self._view_event_handlers))
This closure keeps a reference to self. Deleting those two attributes
releases the reference.

References to *bound* methods were kept in self._event_handlers and
self._view_event_handlers. When the windows closes, clearing those
dict removes several references to self. This was found in both
Win32Window and XlibWindow. I couldn't find anything like that for
CocoaWindow or CarbonWindow.

The last place was in BaseWindow._event_queue. We also clear the list
when closing the window. We have to pay attention to first close the
window through the OS, using the code in the Derived window `close()`
method before calling the Base window `close()` method.