Synchronize audio to master clock.
This is an experiment using DirectSound only. It won't work at all
with other Audio Players.

The audio compensation creates some "scratchy" noises while the
audio is synchronizing to the master clock.

Also the audio player must pass itself down to the source in order
to allow the audio decoding function to know if compensation is
needed or not. This creates an ugly tight coupling.