Refactor and cleanup ffmpeg.py
av.py was the ctypes wrapper around AVbin. Now that we are using
FFmpeg, there are no reasons to have this file. I've moved all the
code that was there to ffmpeg.py and I cleaned up many C Structs
that had become unnecessary.

Moved the functions decode_video and decode_audio into the
FFmpegSource class. This allows to have an audio resample context
and a video rescale context initiated when opening the streams
and re-used each time those functions are called. In the previous
incarnation, those contexts were always re-created and destroyed
at each function call.

More cleanup will be needed, but this is a good step in the right
direction.