Fix SourceGroup
It was not very clear what purpose had SourceGroup. I used it only
to save re-creating the audio player and the texture if we have the
same audio and video format.

The logic about looping as been moved to the Player itself.

An event `on_player_next_source` was added to inform anyone using
the Player that a new source will start playing. It allows client
code to adjust the window size to the video size for instance.