Fix a bug in FFmpeg Source __init__
We were seeking at position 0 when initializing the file. This
allowed source with a start_time > 0 to start at their offset
time.

But a Vorbis audio file showed that when not seeking, the packets
were starting normally at timestamp 0. But when seeking, we were
starting at timestamp 0.3. If the seek was done during the play,
we got the correct packets starting at timestamp 0.

It was decided to fill the queue, and only seek back to 0 if the
source has a start_time > 0. Only for those media (rare) the
first filling will be a bit of a waste.