Improve Media Player seek during mouse drag
On Linux dragging the mouse on the slider was triggering on_mouse_drag events
at a high frequency. This was starving the event loop with seek requests
which take a considerable amount of time (~200ms).
Here we limit the frequency of on_change events triggered by the Slider,
recording the last mouse position when the event is triggered.