On Linux, lib loading should prefer _ld_so_cache
On Linux, ctypes.util.find_library returns the name of the library but not its
entire path. If the library is not on the path search by ld or gcc,
ctypes.cdll.LoadLibrary will NOT find the library.

Pyglet avoid this problem by building a cache with so files. But this is not
used if find_library finds something. Instead we should prefer a fully qualified
path and only if it does not exist, we will try to load what find_library returned.