Fix regression bug in event.py
During event dispatching, when trying to find an event handler on the instance,
we were catching AttributeError as a sign that no handler was found. But there
could be a handler which throws an AttributeError. And this would be silently
swallowed -> bad. Now we check if the instance has an event handler and then
call it.